+++ src/main/kotlin/bridge/domain/GameCommand.kt
@@ -1,4 +1,4 @@
-package bridge
+package bridge.domain
 
 import bridge.sentence.ErrorSentence
